constructor validation moved to utils
