+++ lib/cli/utils.d.ts
@@ -1,2 +1,3 @@
 export declare function executeCommand(command: string, verbose?: boolean): Promise<string>;
 export declare function download(url: string, verbose?: boolean): Promise<void>;
+export declare function validateContract(contractPath: string): Promise<boolean>;
